refactor: move html dict check
