+++ app/shell/py/pie/tests/test_picasso.py
@@ -16,7 +16,7 @@ def test_generate_rule_basic():
         "\t$(Q)process-yaml $< $@\n"
         "build/foo/bar.html: build/foo/bar.md build/foo/bar.yml\n"
         "\t$(Q)$(PANDOC_CMD) $(PANDOC_OPTS) --metadata-file=build/foo/bar.yml -o $@ $<\n"
-        "\t$(Q)detect-html-dicts $@"
+        "\t$(Q)check-bad-jinja-output $@"
     )
     assert rule == expected
